Version 1.1.3

+++ application/controllers/Team.php
@@ -20,9 +20,24 @@ class Team extends CI_Controller {
 	 */
 	public function index()
 	{
-		$this->load->template('team/index','',1);
+		
+		$role_id = 1;
+		
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='team' and me_menu_access.view_access =1";
+        $data["access"] = $this->menumodal->ShowMenuBySearch($condition);
+		if(!empty($data["access"])){
+			$this->load->template('team/index','',1);
+		}
+		else{
+			$this->load->template('error/permission',$data,1);
+		}
 	}
 	public function response(){
+		$role_id = 1;
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='team' and me_menu_access.edit_access =1";
+        $data["edit_access"] = $this->menumodal->ShowMenuBySearch($condition);
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='team' and me_menu_access.delete_access =1";
+        $data["delete_access"] = $this->menumodal->ShowMenuBySearch($condition);
 		$draw = intval($this->input->get("draw"));
 		$start = intval($this->input->get("start"));
 		$length = intval($this->input->get("length"));
@@ -31,12 +46,20 @@ class Team extends CI_Controller {
 		$count=0;     
 		foreach($data["team"] as $r) {
 			$TeamID = $r["TeamID"];
-			$btn = '<button type="button" class="btn btn-soft-primary waves-effect waves-light">';
+			$btn = '';
+			$TeamIDencode = urlencode(base64_encode($TeamID));
+			if(!empty($data["edit_access"])){
+			$btn .= '<a href="'.base_url().'team/edit/'.$TeamIDencode.'" class="btn btn-soft-primary waves-effect waves-light">';
 			$btn .= '<i class="mdi mdi-pencil font-size-16 align-middle" style="line-height: 1;"></i>';
-			$btn .= '</button>';
-			$btn .= '<button onClick="delete_team(this.id)" id="'.$TeamID.'" type="button" class="btn btn-soft-danger waves-effect waves-light">';
-			$btn .= '<i class="bx bx-block font-size-16 align-middle"></i>';
-			$btn .= '</button>';
+			$btn .= '</a>';
+			}
+			if(!empty($data["delete_access"])){
+				$btn .= '<button onClick="delete_team(this.id)" id="'.$TeamID.'" type="button" class="btn btn-soft-danger waves-effect waves-light">';
+				$btn .= '<i class="bx bx-block font-size-16 align-middle"></i>';
+				$btn .= '</button>';
+			}
+
+
 			$data1[] = array(
 				++$count,
 				$r["TeamName"],
@@ -54,8 +77,66 @@ class Team extends CI_Controller {
 		);
 		echo json_encode($result);
 	}
+	public function edit($TeamID){
+		$TeamID = base64_decode(urldecode($TeamID));
+		$role_id = 1;
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='team' and me_menu_access.edit_access =1";
+        $data["access"] = $this->menumodal->ShowMenuBySearch($condition);
+		$array = array("TeamID"=>$TeamID,"me_team.TeamStatus"=>1);
+		$data["team"] = $this->teammodal->ShowTeamBySearch($array);
+		$data["department"] = $this->departmentmodal->ShowDepartment();
+		$data["company"] = $this->companymodal->ShowCompany();
+		if(!empty($data["access"])){
+			$this->load->template('team/edit',$data,1);
+		}
+		else{
+			$this->load->template('error/permission',$data,1);
+		}
+	}
+	public function edit_process(){
+		$TeamStatus = 0;
+		$TeamID = $this->input->post("TeamID");
+		if(isset($_POST["TeamStatus"])){
+			$TeamStatus = 1;
+		}
+		$array = array(
+			"DeparmentID"=>$this->input->post("DeparmentID"),
+			"CompanyID"=>$this->input->post("CompanyID"),
+			"TeamName"=>$this->input->post("TeamName"),
+			"TeamStatus"=>$TeamStatus,
+		);
+		$data["update"] = $this->teammodal->EditTeam($array,$TeamID);
+		if($data["update"]==1){
+			echo "Data Updated";
+		}
+		else{
+			echo "Data Failed";
+		}
+	}
     public function dashboard(){
         echo "Dashboard";
     }
+	public function add(){
+		$role_id = 1;
+		$condition = "me_menu_access.RoleID =" . "'" . $role_id . "' and me_menu.MenuLink='team' and me_menu_access.add_access =1";
+        $data["access"] = $this->menumodal->ShowMenuBySearch($condition);
+		$data["department"] = $this->departmentmodal->ShowDepartment();
+		$data["company"] = $this->companymodal->ShowCompany();
+		if(!empty($data["access"])){
+			$this->load->template('team/add',$data,1);
+		}
+		else{
+			$this->load->template('error/permission',$data,1);
+		}
+	}
+	public function process(){
+		$data["save"] = $this->teammodal->AddTeam($_POST);
+		if($data["save"]==1){
+			echo "Data Saved";
+		}
+		else{
+			echo "Data Failed";
+		}
+	}
 	
 }
